pyproject.toml
[project]
name = "fast_plot"
version = "0.0.1"
dependencies = [
  "numpy",
  "matplotlib",
  "pillow"
]
requires-python = ">= 3.8"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"